Locate appropriate cached image and/or apply filters
When image is modified by filters, resulting image and all
intermediate results will be cached in subdirectory .wall-cache
located in the same directory as original image. Files in
the cache have names of the form

     original-image-name__filter1+filter2:arg2+...

where missing arguments mean they were not supplied on the
command line.

Accordingly, before applying filters to an image we check
whether cache already contains requested file or at least
some of intermediate results to avoid having to start every
time from scratch. We first look for requested file, then keep
removing one filter at a time, until we find cached file that
matches the longest starting subsequence of filters specified
on the command line. Only if nothing is found, we start from
actual input file.

For example, in one hypothetical case, we look for cached
files going down the following list

     1. .wall-cache/wallpaper__dim+blur+tint:17
     2. .wall-cache/wallpaper__dim+blur
     3. .wall-cache/wallpaper__dim
     4. wallpaper

When first file is found in the cache, we reverse direction,
and climb up the very same list until we reach the top and
requested file, generating intermediate results one by one.
For example, if file number 3. was found in cache we would

     blur .wall-cache/wallpaper__dim ->
          .wall-cache/wallpaper__dim+blur, and

     tint .wall-cache/wallpaper__dim+blur ->
          .wall-cache/wallpaper__dim+blur+tint:17