Refactor code into is_unmet_dependency_of()
Unmet dependency is any file that either does not exists, or is
older than it's own dependicy. We call it dependancy since, in
all relevant cases, it is itself a dependancy of some other file
we are trying to generate. Since many files in the cache tend to
be symbolic links, relevant mtime here is that of a link itself,
not the file link points to.