Add cache command
The cache command, if followed by an image recipe adds this image
to cache; if followed by directory, it adds to cache all images as
specified by .recipe files in that directory.

In order to fascilitate this, ensure_recipe_is_cached() is also
changed to simply return non-zero exit code on invalid .recipe file
instead of stopping the entire script. While this behaviour was
appropriate in case of setting a new wallpaper, it shouldn't stop
further processing of other .recipe files.