Add special filename pick:
This filename can be given in place of any directory on the command
line, followed by a list of files, and is replaced by a temporary
directory containing links to these files. It is meant as a simple
way to act on an ad-hoc choice of files.

Note that, at the moment, the only command that accepts directories
is set command; this is likely going to change very soon.