Make browse work on any directory
Rewrite browse to take as optional argument a directory to browse,
defaulting to directory of current wallpaper. We use get_proposed_file
to get pick: ad-hoc file selection and tag queries along with
explicitly specifying directory to browse. In both cases, files are
gathered in $TAG_QUERY_RESULT.browse directory, since
$TAG_QUERY_RESULT might hold current wallpaper and active file
selection. If no wallpaper end up beeing selected, $TAG_QUERY_RESULT.browse
is simply deleted, otherwise it is moved to $TAG_QUERY_RESULT and
wallpaper is again selected in the new location.

Note that "-L" is also added to LS_FLAGS, since most of the files in
$TAG_QUERY_RESULT end up beeing links, and without this flag, they
are sorted according to the mtime of link itself, not file it points to,
producing wrong order during browsing.