Rewrite compose recipe argument
When parsing image recipe, set_command rewrites primary dependency,
the image recipe is starting from, to an absolute path of the image
script actually ends up using (and this image can be different from
the one specified in the recipe due to random shuffling, or tag and
directory expansions).

Before introduction of compose filter, this provided several
advantages:

     - cache1 name contains all the data necessary to reconstruct
       the exact image;

     - different images allways have different cache1 names, which
       prevent flickering when writing over current wallpaper.

In order to keep these advantages, extened this to recipe arguments
of compose filter. Since the goal is to change cache1 name of the
produced image to contain the exact primary dependency used to
generate it, this generation has to happen while this name is
still in construction and move from compose filter to get_recipe().